requirements.txt

gymnasium
numpy<2.0
nes_py==8.2.1
torch<2.2
torchvision
gym-super-mario-bros==7.4.0
tensordict==0.3.0
torchrl==0.3.0
wandb
opencv-python
stable_baselines3
shimmy
tensorboard
torchrl